+++ FOX/__init__.py
@@ -11,6 +11,8 @@ https://auto-fox.readthedocs.io/en/latest/
 
 """
 
+import warnings
+
 from .__version__ import __version__
 
 from .functions import (
@@ -38,7 +40,10 @@ from .ff import (
 
 __author__ = "Bas van Beek"
 __email__ = 'b.f.van.beek@vu.nl'
-example_xyz: str = get_example_xyz()
+
+with warnings.catch_warnings():
+    #: The path+filename of the example multi-xyz file.
+    example_xyz: str = get_example_xyz()
 
 __all__ = [
     '__version__',