+++ CHANGELOG.rst
@@ -8,7 +8,9 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 0.6.9
 *****
-See https://github.com/nlesc-nano/auto-FOX/issues/52.
+* ``FOX.get_example_xyz()`` has been deprecated in favor of ``FOX.example_xyz``.
+
+From https://github.com/nlesc-nano/auto-FOX/issues/52:
 * Repos of script to analyze AMRC data.
 * Simultaneous fitting of different trajectories with different atom types;
   ensure that the PES descriptor generators can have different arguments for each trajectory.